# Conflicts:
#	server/src/main/java/org/elasticsearch/index/shard/IndexShard.java
#	server/src/test/java/org/elasticsearch/index/shard/IndexShardTests.java

+++ test/framework/src/main/java/org/elasticsearch/test/rest/ESRestTestCase.java
@@ -30,7 +30,7 @@ import org.apache.http.entity.StringEntity;
 import org.apache.http.message.BasicHeader;
 import org.apache.http.nio.conn.ssl.SSLIOSessionStrategy;
 import org.apache.http.ssl.SSLContexts;
-import org.apache.lucene.util.IOUtils;
+import org.elasticsearch.core.internal.io.IOUtils;
 import org.elasticsearch.action.admin.cluster.node.tasks.list.ListTasksAction;
 import org.elasticsearch.client.Response;
 import org.elasticsearch.client.ResponseException;
@@ -145,9 +145,11 @@ public abstract class ESRestTestCase extends ESTestCase {
      */
     @After
     public final void cleanUpCluster() throws Exception {
-        wipeCluster();
-        waitForClusterStateUpdatesToFinish();
-        logIfThereAreRunningTasks();
+        if (preserveClusterUponCompletion() == false) {
+            wipeCluster();
+            waitForClusterStateUpdatesToFinish();
+            logIfThereAreRunningTasks();
+        }
     }
 
     @AfterClass
@@ -173,6 +175,17 @@ public abstract class ESRestTestCase extends ESTestCase {
      */
     protected static RestClient adminClient() {
         return adminClient;
+    }
+
+    /**
+     * Returns whether to preserve the state of the cluster upon completion of this test. Defaults to false. If true, overrides the value of
+     * {@link #preserveIndicesUponCompletion()}, {@link #preserveTemplatesUponCompletion()}, {@link #preserveReposUponCompletion()}, and
+     * {@link #preserveSnapshotsUponCompletion()}.
+     *
+     * @return true if the state of the cluster should be preserved
+     */
+    protected boolean preserveClusterUponCompletion() {
+        return false;
     }
 
     /**
@@ -290,7 +303,7 @@ public abstract class ESRestTestCase extends ESTestCase {
 
         if (mustClear) {
             adminClient().performRequest("PUT", "/_cluster/settings", emptyMap(), new StringEntity(
-                clearCommand.string(), ContentType.APPLICATION_JSON));
+                Strings.toString(clearCommand), ContentType.APPLICATION_JSON));
         }
     }
